LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/back-forward-cache/eligibility/inflight-fetch-cors-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: URL is not valid or contains user credentials.
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Invalid URL

Harness Error (TIMEOUT), message = null

TIMEOUT Eligibility (in-flight fetch): CORS succeeded when in BFCache Test timed out
NOTRUN Eligibility (in-flight fetch): CORS failed when in BFCache
